Add files via upload
This is for practice purposes only. Displaying how the dataset in the form of csv can be uploaded and preprocessed. In this pre-processing, missing data is replaced by average/mean values, the categorical data is encoded, the dataset is split into test and training set, and in the end, scaling is performed